+++ config/webpack.package.js
@@ -2,10 +2,10 @@ const webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-// const ngtools = require('@ngtools/webpack');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const commonConfig = require('./webpack.common');
 const { ENV, dir, APP_VERSION } = require('./helpers');
+const { CheckerPlugin } = require('awesome-typescript-loader');
+// const ngtools = require('@ngtools/webpack');
 
 const banner =
 `/**
@@ -23,25 +23,10 @@ module.exports = function(env) {
         {
           test: /\.ts$/,
           loaders: [
-            'awesome-typescript-loader'
+            'awesome-typescript-loader',
+            'angular2-template-loader'
           ],
           exclude: [/\.(spec|e2e|d)\.ts$/]
-        },
-        {
-          test: /\.css/,
-          loader:
-            ExtractTextPlugin.extract({
-              fallbackLoader: 'style-loader',
-              loader:'css-loader?sourceMap'
-            })
-        },
-        {
-          test: /\.scss$/,
-          loader:
-            ExtractTextPlugin.extract({
-              fallbackLoader: 'style-loader',
-              loader: 'css-loader?sourceMap!postcss-loader?sourceMap!sass-loader?sourceMap'
-            })
         }
       ]
     },
@@ -69,10 +54,7 @@ module.exports = function(env) {
       'zone.js/dist/zone': 'zone.js/dist/zone'
     },
     plugins: [
-      new ExtractTextPlugin({
-        filename: '[name].css',
-        allChunks: true
-      }),
+      new CheckerPlugin(),
       new webpack.BannerPlugin({
         banner: banner,
         raw: true,