+++ config/webpack.prod.js
@@ -2,9 +2,9 @@ const webpack = require('webpack');
 const webpackMerge = require('webpack-merge');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const commonConfig = require('./webpack.common');
 const { ENV, dir } = require('./helpers');
+const { CheckerPlugin } = require('awesome-typescript-loader');
 
 module.exports = function(env) {
   return webpackMerge(commonConfig({ env: ENV }), {
@@ -25,33 +25,15 @@ module.exports = function(env) {
         {
           test: /\.ts$/,
           loaders: [
-            'awesome-typescript-loader'
-          ],
+            'awesome-typescript-loader',
+            'angular2-template-loader'
+         ],
           exclude: [/\.(spec|e2e|d)\.ts$/]
-        },
-        {
-          test: /\.css/,
-          loader:
-            ExtractTextPlugin.extract({
-              fallbackLoader: 'style-loader',
-              loader:'css-loader?sourceMap'
-            })
-        },
-        {
-          test: /\.scss$/,
-          loader:
-            ExtractTextPlugin.extract({
-              fallbackLoader: 'style-loader',
-              loader: 'css-loader?sourceMap!postcss-loader?sourceMap!sass-loader?sourceMap'
-            })
         }
       ]
     },
     plugins: [
-      new ExtractTextPlugin({
-        filename: '[name].css',
-        allChunks: true
-      }),
+      new CheckerPlugin(),
       new webpack.optimize.CommonsChunkPlugin({
         name: ['polyfills'],
         minChunks: Infinity