+++ config/webpack.test.js
@@ -3,10 +3,15 @@ const webpackMerge = require('webpack-merge');
 const chalk = require('chalk');
 const commonConfig = require('./webpack.common');
 const { ENV, dir } = require('./helpers');
+const combineLoaders = require('webpack-combine-loaders');
+const { CheckerPlugin } = require('awesome-typescript-loader');
 
 module.exports = function(env) {
   return webpackMerge(commonConfig({ env: ENV }), {
     devtool: 'inline-source-map',
+    plugins: [
+      new CheckerPlugin()
+    ],
     module: {
       exprContextCritical: false,
       rules: [
@@ -18,14 +23,28 @@ module.exports = function(env) {
         },
         {
           test: /\.ts$/,
-          loader: 'awesome-typescript-loader',
-          query: {
-            sourceMap: false,
-            inlineSourceMap: true,
-            compilerOptions: {
-              removeComments: true
+          loader: combineLoaders([
+            {
+              loader: 'awesome-typescript-loader',
+              query: {
+                sourceMap: false,
+                inlineSourceMap: true,
+                compilerOptions: {
+                  removeComments: true
+                }
+              }
+            },
+            {
+              loader: 'angular2-template-loader',
+              query: {
+                sourceMap: false,
+                inlineSourceMap: true,
+                compilerOptions: {
+                  removeComments: true
+                }
+              }
             }
-          },
+          ]),
           exclude: [/\.e2e\.ts$/, /(node_modules)/]
         },
         {
@@ -37,14 +56,6 @@ module.exports = function(env) {
             /\.(e2e|spec)\.ts$/,
             /node_modules/
           ]
-        },
-        {
-          test: /\.css/,
-          loader: 'style-loader!css-loader?sourceMap'
-        },
-        {
-          test: /\.scss$/,
-          loader: 'style-loader!css-loader!postcss-loader?sourceMap!sass-loader?sourceMap'
         }
       ]
     },