+++ demo/bootstrap.ts
@@ -1,13 +1,8 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { bootloader } from '@angularclass/hmr';
-
 import { AppModule } from './module';
 
-export function main(): Promise<any> {
-  return platformBrowserDynamic()
+ document.addEventListener('DOMContentLoaded', () => {
+   platformBrowserDynamic()
     .bootstrapModule(AppModule)
     .catch(err => console.error(err));
-}
-
-if(HMR) bootloader(main);
-if(!HMR) main();
+ });