+++ demo/module.ts
@@ -1,11 +1,8 @@
 import { NgModule, ApplicationRef } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { removeNgStyles, createNewHosts } from '@angularclass/hmr';
 
 import { NgxDatatableModule } from '../src';
 import { AppComponent } from './app.component';
-import '../src/components/datatable.scss';
-import '../src/themes/material.scss';
 
 // -- Basic
 import { BasicFixedComponent } from './basic/basic-fixed';
@@ -88,18 +85,4 @@ import { ColumnPinningComponent } from './columns/pinning';
   imports: [BrowserModule, NgxDatatableModule],
   bootstrap: [AppComponent]
 })
-export class AppModule {
-
-  constructor(private appRef: ApplicationRef) { }
-
-  hmrOnDestroy(store) {
-    const cmpLocation = this.appRef.components.map(cmp => cmp.location.nativeElement);
-    store.disposeOldHosts = createNewHosts(cmpLocation);
-    removeNgStyles();
-  }
-
-  hmrAfterDestroy(store) {
-    store.disposeOldHosts();
-    delete store.disposeOldHosts;
-  }
-}
+export class AppModule { }