+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Master
+- BREAKING: Renamed `datatable` css class to `ngx-datatable`
+- BREAKING: Embedded core css file with Angular `styleUrls`
+- Enhancement: Allow sorting from inline templates (#431)
+
 ## 5.0.0
 - BREAKING: Re-worked all inputs, methods and outputs of row detail to its own container
 - Enhancement: Implement Finder/Explorer shift selection behavior